Nighttime Progress Report
Created clientConsole as substitute to clientUI. This is to cut down UI
development time.

Client service partly touched and render instruction has been changed.
It is advisable to re-write the entire code as the last one is very
buggy.

As an idea it seems a nice idea to break the renderer into different
module therefore each renderer can be treated differently considering
they act differently.